fix(cloudrun): add region tag and move handler to main
